Shapoorji Hadapsar Annexe Manjri Pune
Shapoorji Pallonji Hadapsar Annexe is a residential project in Manjri, Pune. It is one of the best locations for residential property near NH-65 (Pune-Solapur Road). The project lets you enjoy 1/2/3 BHK residences and includes elegant high-rise exclusive towers. Each of the towers in the project fulfils the needs with the best comfortable and stylish homes. The project has spacious homes with a huge number of luxurious amenities. It comes as an 80% open space development with a beautifully themed classical garden.  

Shapoorji Pallonji Hadapsar Annexe Manjri has some of the best features. The project is well-designed by an internationally renowned designer. It is an under-construction development that is designed with the concept of classical inputs within the building. It comes as a blend of contemporary design and features. The project has a double-height lounge in each of the towers. It has beautifully theme landscapes along with water bodies. The project comes with high-speed elevators and secured entrance gates. The project comes under the surveillance of CCTV surveillance.

Shapoorji Joyville Hadapsar offers some unique amenities like an amphitheatre, a fragrance garden, a central garden, a kid's playground, pavilions, a senior citizen corner, and yoga deck, and a gazebo further boosting its reputation. Here, you can get excellent outdoor sports facilities which include a jogging track, lawn tennis, badminton, squash court, basketball, and skating rink adding value to your comfort.

Shapoorji Pallonji is a 15 decades old real estate organization fully committed to building new landmarks. Lending on its futuristic ideas, ingenious designs, aesthetics, and cutting-edge techs. The developer group succeed to raise the level of people's living standards. Shapoorji Joyville Hadapsar Manjri is one of its kind key residential solution which brings you and your loved ones to the close vicinity of nature. For those looking for an elevated lifestyle or owning a home for investment purposes, Shapoorji Joyville Hadapsar Pune is an ideal choice.

To know more - https://bit.ly/3i4bCnj